requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
sqlmodel==0.0.24
SQLAlchemy==2.0.42
alembic==1.13.2
Mako==1.3.5

python-dotenv==1.0.1

httpx==0.28.1
typing_extensions==4.13.2
mysql-connector-python==9.3.0
pytest==8.3.4
pytest-asyncio==0.24.0

# User-Agent parsing for richer request logs
user-agents==2.2.0

